+++ 1.21/Common/src/main/java/fuzs/easymagic/data/client/ModLanguageProvider.java
@@ -20,5 +20,6 @@ public class ModLanguageProvider extends AbstractLanguageProvider {
         builder.add(BlockConversionHandler.INVALID_BLOCK_COMPONENT, "Unable to open. Break and replace to use.");
         builder.add(ModRegistry.ENCHANTING_CATALYSTS_ITEM_TAG, "Enchanting Catalysts");
         builder.add(ModRegistry.REROLL_CATALYSTS_ITEM_TAG, "Reroll Catalysts");
+        builder.add(ModRegistry.UNALTERED_ENCHANTING_TABLES_BLOCK_TAG, "Unaltered Enchanting Tables");
     }
 }