+++ 1.21/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v21.0.3-1.21] - 2024-09-03
+### Changed
+- Blocks created during world generation are now automatically converted to the mod replacement
+- Some enchanting tables added by other mods are now supported
+
 ## [v21.0.2-1.21] - 2024-07-30
 ### Fixed
 - Fix option for keeping the vanilla enchanting book rendering in the enchanting screen not having any effect